moved more files around so they are in better folders
